rename entity -> thing; rename require_can_write -> require_api_permissions
