+++ openlibrary/asgi_app.py
@@ -218,7 +218,6 @@ def create_app() -> FastAPI | None:
     from openlibrary.fastapi.books import router as books_router
     from openlibrary.fastapi.cdn import router as cdn_router
     from openlibrary.fastapi.checkins import router as checkins_router
-    from openlibrary.fastapi.entities import router as entities_router
     from openlibrary.fastapi.importapi import router as importapi_router
     from openlibrary.fastapi.internal.api import router as internal_router
     from openlibrary.fastapi.languages import router as languages_router
@@ -229,6 +228,7 @@ def create_app() -> FastAPI | None:
     from openlibrary.fastapi.publishers import router as publishers_router
     from openlibrary.fastapi.search import router as search_router
     from openlibrary.fastapi.subjects import router as subjects_router
+    from openlibrary.fastapi.things import router as things_router
     from openlibrary.fastapi.yearly_reading_goals import (
         router as yearly_reading_goals_router,
     )
@@ -242,7 +242,7 @@ def create_app() -> FastAPI | None:
     app.include_router(internal_router)
     app.include_router(languages_router)
     app.include_router(lists_router)
-    app.include_router(entities_router)
+    app.include_router(things_router)
     app.include_router(merge_authors_router)
     app.include_router(partials_router)
     app.include_router(public_my_books_router)
